requirements.txt
bokeh==1.4.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
packaging==20.3
pandas==1.0.1
parse==1.15.0
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
six==1.14.0
tornado==6.0.4
